+++ services-js/access-boston/src/client/preferred-chosen-name/views/errorView.tsx
@@ -3,8 +3,6 @@
 import { jsx, css } from '@emotion/core';
 import { MouseEvent } from 'react';
 
-// import { CommonAttributes } from '../types';
-
 //--- HTML Struct & Styling ---//
 import QuestionComponent from '../components/QuestionComponent';
 import AlertComponent from '../components/AlertComponent';
@@ -12,44 +10,69 @@ import { PREFERRED_NAME_STYLING } from '../styling/index';
 
 interface ErrorProps {
   handleQuit: (ev: MouseEvent) => void;
+  appTitle: string;
 }
 
-export default function ErrorView(props: ErrorProps) {
-  const { handleQuit } = props;
+export const ErrorView2 = (_props: ErrorProps) => {
+  const errorTag = () => {
+    return (
+      <>
+        An error occurred while processing your request, go the{' '}
+        <a
+          target="_blank"
+          href="https://www.boston.gov/departments/innovation-and-technology/access-boston-portal-help"
+        >
+          Access Boston Help
+        </a>{' '}
+        page for more information
+      </>
+    );
+  };
 
   return (
     <div css={PREFERRED_NAME_STYLING}>
-      <div className="BorderedAppWrapper">
-        <div className="AppInnerContainer">
-          <AlertComponent type={`Error`} text="Error Status" />
-          <QuestionComponent
-            quitBtn={true}
-            quitBtnText="Close"
-            handleQuit={handleQuit}
-          >
-            <div className="row" css={BODY_TEXT_STYLING}>
-              <div className="bodyText">
-                <p css={BODY_PARAGRAPH_STYLING}>
-                  There was an error with this request, please try again later.
-                </p>
+      <div className={'AddBorderTop'}>
+        <div className="BorderedAppWrapper">
+          <div className="AppInnerContainer">
+            <AlertComponent
+              type={`Error`}
+              text="Something Went Wrong"
+              altText={errorTag()}
+            />
+
+            <QuestionComponent quitBtn={true} quitBtnText="Close">
+              <div className="row" css={BODY_TEXT_STYLING}>
+                <div className="bodyText">
+                  <button
+                    onClick={() =>
+                      window.open(
+                        'https://www.boston.gov/departments/innovation-and-technology/access-boston-portal-help',
+                        '_blank'
+                      )
+                    }
+                    className={`btn`}
+                  >
+                    Help Page
+                  </button>
+                </div>
               </div>
-            </div>
-          </QuestionComponent>
+            </QuestionComponent>
+          </div>
         </div>
       </div>
     </div>
   );
-}
+};
+
+export default ErrorView2;
 
 const BODY_TEXT_STYLING = css({
-  paddingTop: '10px !important',
-  '@media (max-width: 600px)': {
-    padding: '0 20px 10px !important',
-  },
-});
+  display: 'flex',
+  justifyContent: 'center',
 
-const BODY_PARAGRAPH_STYLING = css({
-  '@media (max-width: 600px)': {
-    lineHeight: '1.5rem  !important',
+  '.btn': {
+    height: '48px',
+    padding: '20px',
+    paddingBottom: '32px',
   },
 });
