+++ services-js/access-boston/src/client/preferred-chosen-name/views/welcomeView.tsx
@@ -1,45 +1,45 @@
 /** @jsx jsx */
 import { jsx, css } from '@emotion/core';
 
-import { MouseEvent } from 'react';
+import { MouseEvent, KeyboardEvent } from 'react';
+import { MEDIA_SMALL_MAX } from '@cityofboston/react-fleet';
 import QuestionComponent from '../components/QuestionComponent';
 
 import { CommonAttributes } from '../types';
-import { PREFERRED_NAME_STYLING } from '../styling/index';
+import { PREFERRED_NAME_STYLING, WELCOMEVIEW_STYLING } from '../styling/index';
 
 interface welcomeProps {
-  handleProceed: (ev: MouseEvent) => void;
+  handleProceed: (ev: MouseEvent | KeyboardEvent) => void;
   appTitle: string;
   state: CommonAttributes;
 }
 
-export default function WelcomeView({ handleProceed, state }: welcomeProps) {
-  // console.log(`state: `, state);
-
+export const WelcomeView2 = ({ handleProceed, state }: welcomeProps) => {
   return (
     <div css={PREFERRED_NAME_STYLING}>
+      <h2 className="headerBlock">Update Chosen Name</h2>
+
       <div className="BorderedAppWrapper">
         <div className="AppInnerContainer">
-          <h2 className="headerBlock" css={HEADER_STYLING} />
           <QuestionComponent
             quitBtn={false}
             nextButtonText="Continue"
             allowProceed={true}
             handleProceed={handleProceed}
           >
-            <div css={SNIPPET_CONTAINER_STYLING}>
+            <div css={WELCOMEVIEW_STYLING}>
               <p>
                 A chosen name could be a preferred nickname, middle name, a
                 shortened version of a legal name, or a name that aligns with
                 your gender identity or expression.
               </p>
+
+              <label>What’s Changing:</label>
               <ul>
-                <strong>What’s Changing: </strong>
                 <li>
                   <strong>Chosen Name: </strong>
                   You can add a chosen name to your profile. This name will be
-                  displayed in internal City of Boston systems and
-                  communications.
+                  displayed in internal City of Boston systems.
                 </li>
                 {!state.altWorkflow && (
                   <li>
@@ -49,21 +49,24 @@ export default function WelcomeView({ handleProceed, state }: welcomeProps) {
                   </li>
                 )}
               </ul>
+
+              <label>What’s Not Changing: </label>
               <ul>
-                <strong>What’s Not Changing: </strong>
                 {state.altWorkflow && (
                   <li>
                     <strong>Email Address: </strong>Your email address will
-                    remain the same.
+                    remain the same. To update your email, please contact your
+                    IT Department.
                   </li>
                 )}
                 <li>
                   <strong>Legal Name: </strong>
                   City Employees, your legal name used for tax documents (like
                   W2s), paystubs, or other official legal documents will remain
-                  unchanged. To change your legal name, please follow the{' '}
+                  unchanged. This will not update HCM or ESS. To change your
+                  legal name, please follow the{' '}
                   <a
-                    href="https://sso-test.boston.gov/as/authorization.oauth2?response_type=code&client_id=pa_wam&redirect_uri=https%3A%2F%2Fess-awsuat.boston.gov%2Fpa%2Foidc%2Fcb&state=eyJ6aXAiOiJERUYiLCJhbGciOiJkaXIiLCJlbmMiOiJBMTI4Q0JDLUhTMjU2Iiwia2lkIjoiQk9MZmJmTm1YZFBRdHhsTkY4aEdUT0hQa1NnIiwic3VmZml4IjoiMVhyR2tVLjE3MzI4OTA5MzMifQ..HK2KIBY6u5vJFcsaXyv9NA.RDqP6cAoD_aRK_dOL6t4w46vt8L-YaLpo5LGPkI1C6_nUIE7cAAnd-AF_N2KATyqOyx1hr6nhUgVuX3TfjX1xM6bc7_4_0c2XUnxPCN9700.IxU5tLkqs_LjSdGlQx_GaA&nonce=nZdehsu67Ivv9NoymGf8FQiQlxfsCmbNdBpyDOqeSZQ&scope=openid%20profile%20email&vnd_pi_requested_resource=https%3A%2F%2Fess-awsuat.boston.gov%2F&vnd_pi_application_name=ESS+UAT+AWS+Cloud+application+"
+                    href="https://ess.boston.gov/pages/NameChange-BS3/namechange.aspx"
                     target="_blank"
                   >
                     legal name change
@@ -71,10 +74,10 @@ export default function WelcomeView({ handleProceed, state }: welcomeProps) {
                   .
                 </li>
               </ul>
-              <p>
+              <p css={PLEASE_NOTE__STYLING}>
                 <strong>Please note: </strong>
                 Any changes made to your chosen name or email address will be
-                reflected across all relevant systems and communications.
+                reflected across all relevant systems.
               </p>
             </div>
           </QuestionComponent>
@@ -82,53 +85,13 @@ export default function WelcomeView({ handleProceed, state }: welcomeProps) {
       </div>
     </div>
   );
-}
+};
 
-const SNIPPET_CONTAINER_STYLING = css({
-  lineHeight: '2rem !important',
-  padding: '0 40px',
-  '& > p, & > ul': {
-    borderBottom: '1px solid #ccc',
-    padding: '20px 0px',
-    margin: '0px',
-    '@media (max-width: 600px)': {
-      lineHeight: '1.5rem  !important',
-      padding: '25px 0px',
-    },
-  },
-  '& > p:last-of-type': {
-    borderBottom: 'none',
-  },
-  '& ul': {
-    '& > strong': {
-      display: 'block',
-      marginBottom: '25px',
-      '@media (max-width: 600px)': {
-        marginBottom: '10px',
-      },
-    },
-  },
-  '& li': {
-    marginLeft: '40px',
-    marginBottom: '20px',
-    listStyleType: 'disc',
-    listStyle: 'outside',
-    '@media (max-width: 600px)': {
-      marginLeft: '20px',
-      marginBottom: '10px',
-    },
-  },
-  '@media (max-width: 600px)': {
-    padding: '0px 15px',
-  },
-});
+export default WelcomeView2;
 
-// Add CSS for changing header text on mobile
-const HEADER_STYLING = css({
-  '&::before': {
-    content: '"Update Chosen Name"',
-    '@media (max-width: 600px)': {
-      content: '"Update Chosen Name"',
-    },
+const PLEASE_NOTE__STYLING = css({
+  [MEDIA_SMALL_MAX]: {
+    paddingLeft: '22px',
+    paddingRight: '22px',
   },
 });
